Split things in different files
It makes more sense like that.
